Set C bindings for unix_readdir* C bindings in their own files
