+++ src/unix/lwt_unix_unix.h
@@ -44,147 +44,6 @@
 #include "unix_recv_send_utils.h"
 
 /* +-----------------------------------------------------------------+
-   | JOB: rewinddir                                                  |
-   +-----------------------------------------------------------------+ */
-
-
-
-/* +-----------------------------------------------------------------+
-   | JOB: readdir                                                    |
-   +-----------------------------------------------------------------+ */
-
-struct job_readdir {
-    struct lwt_unix_job job;
-    DIR *dir;
-    struct dirent *entry;
-    int error_code;
-};
-
-static void worker_readdir(struct job_readdir *job)
-{
-    // From the man page of readdir
-    // If the end of the directory stream is reached, NULL is returned and
-    // errno is not changed. If an error occurs, NULL is returned and errno
-    // is set appropriately. To distinguish end of stream and from an error,
-    // set errno to zero before calling readdir() and then check the value of
-    // errno if NULL is returned.
-    errno = 0;
-    job->entry = readdir(job->dir);
-    job->error_code = errno;
-}
-
-static value result_readdir(struct job_readdir *job)
-{
-    LWT_UNIX_CHECK_JOB(job, job->entry == NULL && job->error_code != 0,
-                       "readdir");
-    if (job->entry == NULL) {
-        // From the man page
-        // On success, readdir() returns a pointer to a dirent structure.
-        // (This structure may be statically allocated; do not attempt to
-        // free(3) it.)
-        lwt_unix_free_job(&job->job);
-        caml_raise_end_of_file();
-    } else {
-        value name = caml_copy_string(job->entry->d_name);
-        // see above about not freeing dirent
-        lwt_unix_free_job(&job->job);
-        return name;
-    }
-}
-
-CAMLprim value lwt_unix_readdir_job(value val_dir)
-{
-    LWT_UNIX_INIT_JOB(job, readdir, 0);
-    job->dir = DIR_Val(val_dir);
-    return lwt_unix_alloc_job(&job->job);
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: readdir_n                                                  |
-   +-----------------------------------------------------------------+ */
-
-struct job_readdir_n {
-    struct lwt_unix_job job;
-    DIR *dir;
-    /* count serves two purpose:
-       1. Transmit the maximum number of entries requested by the programmer.
-          See `readdir_n`'s OCaml side documentation.
-       2. Transmit the number of actually read entries from the `worker` to
-          the result parser.
-          See below `worker_readdir_n` and `result_readdir_n`
-    */
-    long count;
-    int error_code;
-    char *entries[];
-};
-
-static void worker_readdir_n(struct job_readdir_n *job)
-{
-    long i;
-    for (i = 0; i < job->count; i++) {
-        errno = 0;
-        struct dirent *entry = readdir(job->dir);
-
-        /* An error happened. */
-        if (entry == NULL && errno != 0) {
-            job->count = i;
-            job->error_code = errno;
-            return;
-        }
-
-        /* End of directory reached */
-        if (entry == NULL && errno == 0) break;
-
-        /* readdir is good */
-        char *name = strdup(entry->d_name);
-        if (name == NULL) {
-            job->count = i;
-            job->error_code = errno;
-            return;
-        }
-
-        /* All is good */
-        job->entries[i] = name;
-    }
-    job->count = i;
-    job->error_code = 0;
-}
-
-static value result_readdir_n(struct job_readdir_n *job)
-{
-    CAMLparam0();
-    CAMLlocal1(result);
-    int error_code = job->error_code;
-    if (error_code) {
-        long i;
-        for (i = 0; i < job->count; i++) free(job->entries[i]);
-        lwt_unix_free_job(&job->job);
-        unix_error(error_code, "readdir", Nothing);
-    } else {
-        result = caml_alloc(job->count, 0);
-        long i;
-        for (i = 0; i < job->count; i++) {
-            Store_field(result, i, caml_copy_string(job->entries[i]));
-        }
-        for (i = 0; i < job->count; i++) free(job->entries[i]);
-        lwt_unix_free_job(&job->job);
-        CAMLreturn(result);
-    }
-}
-
-CAMLprim value lwt_unix_readdir_n_job(value val_dir, value val_count)
-{
-    long count = Long_val(val_count);
-    DIR *dir = DIR_Val(val_dir);
-
-    LWT_UNIX_INIT_JOB(job, readdir_n, sizeof(char *) * count);
-    job->dir = dir;
-    job->count = count;
-
-    return lwt_unix_alloc_job(&job->job);
-}
-
-/* +-----------------------------------------------------------------+
    | JOB: readlink                                                   |
    +-----------------------------------------------------------------+ */
